Major update of all the things.

+++ Busidex.Utils/Converters/ValueProgressBarConverter.cs
@@ -7,7 +7,7 @@ namespace Busidex.Utils.Converters
     {
         public object Convert(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
         {
-            return System.Convert.ToDouble(value) / 100;
+            return System.Convert.ToDecimal(value) / 100;
         }
         public object ConvertBack(object value, Type targetType, object parameter, System.Globalization.CultureInfo culture)
         {
